==> 01-process-images.Rmd <==
---
title: "Image Preprocessing"
author: "Kelly Hondula"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(raster)
library(readr)
library(fs)
library(stringr)
library(dplyr)
library(colorspace)
library(sf)
library(binaryLogic)
library(sp)
library(velox)
library(readr)
library(dplyr)
```

# Apply UDM

* The Unusable data mask from Planet described [here](https://github.com/planetlabs/notebooks/blob/master/jupyter-notebooks/udm/udm.ipynb)
* Most conservative - only keep pixels where the value of the UDM is 0, i.e. there are no flags. 
* UDMs are saved in `metadata/download_udm/

```{r check_udms}
myimg_ids <- substr(list.files("data/data/"), 1, 20) # 20 character image IDs
message(paste(length(myimg_ids), "downloaded SR images"))
my_udms <- substr(list.files("data/metadata/download_udm"), 1, 20)
message(paste(length(my_udms), "UDMs"))
message(paste(sum(!myimg_ids %in% my_udms), "images without UDMs"))
```


```{r, eval = FALSE}
imgs <- data.frame("img_id_20" = myimg_ids,
           "full_id" = basename(list.files('data/data')),
           stringsAsFactors = FALSE)
udms <- data.frame("udm_id_20" = my_udms,
                   "udm_full_id" = basename(list.files('data/metadata/download_udm')),
                   stringsAsFactors = FALSE)
imgs_masked <-  data.frame("udm_masked" =
                      tools::file_path_sans_ext(basename(
                        list.files('data/udm_masked_imgs/', pattern = ".gri"))), 
                      stringsAsFactors = FALSE)
imgs_bands_added <- data.frame("bands_added" =
                      tools::file_path_sans_ext(basename(list.files('data/imgs_for_analysis/'))), 
                      stringsAsFactors = FALSE)


imgs %>% 
  full_join(udms, by = c('img_id_20' = 'udm_id_20')) %>%
  mutate(udm_masked = img_id_20 %in% imgs_masked$udm_masked,
         bands_added = img_id_20 %in% imgs_bands_added$bands_added) %>% 
  arrange(img_id_20) %>% 
  readr::write_csv('imgs_inventory.csv')
```

Mask the image to set all values as NA that are not 0 in the unusable data mask

First make a vector of the images that are in data/ but not in udm_masked_imgs

```{r ids_to_mask}
all_masked_imgs <- substr(list.files("data/udm_masked_imgs", pattern = ".grd"), 1, 20)
message(paste(length(all_masked_imgs), "udm masked images"))
imgIDs_to_mask <- myimg_ids[!myimg_ids %in% all_masked_imgs]
message(paste(length(imgIDs_to_mask), "udm not applied"))
message("Not masked: ", paste(imgIDs_to_mask, collapse = ", "))
imgIDs_to_mask_w_udms <- imgIDs_to_mask[imgIDs_to_mask %in% my_udms]
message(paste(length(imgIDs_to_mask_w_udms), "have udm but not applied"))
imgIDs_to_mask_w_udms
```

Apply UDM mask to vector you just created - IDs with UDMs that arent in udm_masked_imgs folder:

```{r udm_func_defn}
# imgID <- imgIDs_to_mask_w_udms[1]

apply_udm_mask <- function(imgID){
  myimg <- stack(grep(imgID, 
                   list.files("/nfs/khondula-data/planetmethane/data/", 
                              full.names = TRUE), value = TRUE))
  myimg_udm <- raster(grep(imgID, 
                   list.files("/nfs/khondula-data/planetmethane/metadata/download_udm", 
                              full.names = TRUE), value = TRUE))

  masked_filename = paste0("/nfs/khondula-data/planetmethane/udm_masked_imgs/", imgID, ".grd")
  masked_stack <- mask(myimg, myimg_udm, inverse = TRUE, maskvalue = 0, updatevalue = NA,
                     filename = masked_filename, overwrite = TRUE)  
}

```

Run function locally 

```{r}
apply_udm_mask(imgID = imgIDs_to_mask_w_udms[1])
# purrr::walk(imgIDs_to_mask_w_udms[], ~apply_udm_mask(.x))
```

Run function On the cluster

```{r}
library(rslurm)
pars <- data.frame(imgID = imgIDs_to_mask_w_udms, stringsAsFactors = FALSE)

sjob <- rslurm::slurm_apply(apply_udm_mask, pars, 
                            jobname = "udms",
                            slurm_options = list(time = "12:00:00", partition = "sesync"),
                            nodes = 5, 
                            cpus_per_node = 2, 
                            submit = TRUE)

print_job_status(sjob)
# rslurm::cancel_slurm(sjob)
```

# Add Bands

First figure out which images will be used for analysis and need the additional bands added

```{r}
masked_imgfiles <- fs::dir_ls("data/udm_masked_imgs/", glob = "*.grd")
masked_ids <- substr(basename(masked_imgfiles), 1, 20)
message(paste(length(masked_ids), "udm masked images"))

addbands_files <- fs::dir_ls("data/imgs_for_analysis")
addbands_ids <- substr(basename(addbands_files), 1, 20)
message(paste(length(addbands_ids), "images for analysis with bands added"))

# needbands_ids <- names(which(masked_layers)) %>% basename() %>% substr(1,20)
# 
needbands_files <- masked_imgfiles[!masked_ids %in% addbands_ids]

needbands_ids <- substr(basename(needbands_files), 1, 20)
needbands_ids
message(paste(length(needbands_ids), "udm masked images without bands added"))

```

Update image metadata file! 

```{r}
img_metadata_df <- readr::read_csv("data/img_metadata_df.csv")
imgs_addbands <- img_metadata_df %>%
  filter(cloud_cover >=0.01 & cloud_cover < 0.1) %>%
  arrange(desc(cloud_cover)) %>%
  pull(imgID)

imgs_metadata_issues <- img_metadata_df %>%
  dplyr::filter(quality_category == "test" | cloud_cover > 0.1 | visible_clouds) %>% 
  pull(imgID)

needbands_files <- needbands_files[!needbands_ids %in% imgs_metadata_issues]
needbands_ids <- substr(basename(needbands_files), 1, 20)
needbands_ids
message(paste(length(needbands_ids), "masked images without bands added and without metadata issues"))
```


```{r addbands_func_defn}
masked_imgfile <- needbands_files[1]
# masked_imgfile <- masked_imgfiles[1]

save_allbands <- function(masked_imgfile){
  # read in original 4 bands
  r_stack <- raster::stack(masked_imgfile)
  # get imgID
  imgID <- substr(basename(masked_imgfile), 1, 20)
  
  img_metadata_df <- readr::read_csv("/nfs/khondula-data/planetmethane/img_metadata_df.csv")
  imgs_metadata_issues <- img_metadata_df %>% 
  dplyr::filter(quality_category == "test" | cloud_cover > 0.1 | visible_clouds) %>% pull(imgID)

  if(!imgID %in% imgs_metadata_issues){
    
  new_filename <- sprintf("/nfs/khondula-data/planetmethane/imgs_for_analysis/%s.tiff", imgID)
  # add NDVI layer
  r_stack$NDVI <- raster::overlay(r_stack[[3]], r_stack[[4]],
                                   fun = function(b3, b4){return((b4-b3)/(b4+b3))})
                                   # filename = new_filename)
  # add NDWI layer
  r_stack$NDWI <- raster::overlay(r_stack[[2]], r_stack[[4]],
                                   fun = function(b2, b4){return((b2-b4)/(b2+b4))})
                                   # filename = new_filename)
  # get vectors of band values
  b1_blue <- velox::velox(r_stack)$as.matrix(band = 1) %>% as.vector()
  b2_green <- velox::velox(r_stack)$as.matrix(band = 2) %>% as.vector()
  b3_red <- velox::velox(r_stack)$as.matrix(band = 3) %>% as.vector()
  b4_NIR <- velox::velox(r_stack)$as.matrix(band = 4) %>% as.vector()
  
  # convert to data frame
  rgbvals <- data.frame("b1_blue" = b1_blue, 
                        "b2_green" = b2_green, 
                        "b3_red" = b3_red) %>%
    dplyr::mutate(cellid = 1:ncell(r_stack)) 
  
  rgbvals_narm <- rgbvals %>% dplyr::filter(!is.na(b1_blue), !is.na(b2_green), !is.na(b3_red))
  # calculate hue saturation value
  hsvvals <- rgb2hsv(r = rgbvals_narm[,"b3_red"],
                   g = rgbvals_narm[,"b2_green"], 
                   b = rgbvals_narm[,"b1_blue"], 
                   maxColorValue = max(rgbvals_narm[,1:3]))
  # calculate chroma luminance hue2
  labvals <- colorspace::RGB(rgbvals_narm[,"b3_red"], 
                           rgbvals_narm[,"b2_green"],
                           rgbvals_narm[,"b1_blue"]) %>% as("polarLAB")
  
  # get all values together in a dataframe
  hsv_df <- hsvvals %>% t() %>% as.data.frame() %>% 
    dplyr::mutate(cellid = rgbvals_narm$cellid)

  labvals_df <- labvals@coords %>% as.data.frame() %>% 
    dplyr::mutate(cellid = rgbvals_narm$cellid)
  
  colmetrics_df <- rgbvals %>% 
    dplyr::left_join(hsv_df) %>% 
    dplyr::left_join(labvals_df)

  # define new function to make raster layer from column
  column_to_rasterlayer <- function(colorcol){
    lum_mat <- matrix(colmetrics_df[[colorcol]], nrow(r_stack), ncol(r_stack))
    lum_vx <- velox::velox(lum_mat, extent = extent(r_stack), res = res(r_stack))
    return(lum_vx$as.RasterLayer())}

  r_stack$HUE <- column_to_rasterlayer("h")
  r_stack$SAT <- column_to_rasterlayer("s")
  r_stack$VAL <- column_to_rasterlayer("v")
  r_stack$LUM <- column_to_rasterlayer("L")
  r_stack$CHROMA <- column_to_rasterlayer("C")
  r_stack$HUE2 <- column_to_rasterlayer("H")

  raster::writeRaster(r_stack, new_filename, overwrite = TRUE)    
  }

  }
```

```{r}
save_allbands(masked_imgfile = masked_imgfiles[3])
purrr::walk(needbands_files, ~save_allbands(masked_imgfile = .x))
```

```{r}
library(rslurm)
# masked_imgfiles[7]
needbands_files_nfs <- gsub("data", "/nfs/khondula-data/planetmethane", 
                            needbands_files)

pars <- data.frame(masked_imgfile = needbands_files_nfs, stringsAsFactors = FALSE)

sjob4 <- rslurm::slurm_apply(save_allbands, pars, 
                            jobname = "addbands",
                            pkgs = c("raster", "readr", "dplyr", "velox", "colorspace"),
                            slurm_options = list(partition = "sesync", 
                                                 time = "12:00:00"),
                            nodes = 8, 
                            cpus_per_node = 2, 
                            submit = TRUE)

print_job_status(sjob4)

```


```{r}
length(list.files("data/imgs_for_analysis"))
# band_names <- c("b1_blue", "b2_green", "b3_red", "b4_NIR",
#                 "NDVI", "NDWI", "HUE", "SAT", "VAL", "LUM", "CHROMA", "HUE2")
# r_stack_in <- raster::stack("data/imgs_for_analysis/20171004_155646_103f.tif")
# names(r_stack_in) <- band_names
# tabularaster::as_tibble(r_stack_in) %>%
#   dplyr::filter(!is.na(cellvalue)) %>%
#   head()
```
