==> 02-prep-training.Rmd <==
---
title: "Prepare training data"
author: "Kelly Hondula"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(sf)
library(dplyr)
library(fasterize)
library(leaflet)
library(readr)
library(raster)
library(stringr)
data_dir <- "data"
```

# Create AOI

This creates AOI and AOI_union polygons

Download watershed boundary for 

* Tuckahoe Creek 0206000501
* Watts Creek-Choptank 0206000502

```{r}
download.file("https://prd-tnm.s3.amazonaws.com/StagedProducts/Hydrography/NHD/HU8/HighResolution/Shape/NHD_H_02060005_HU8_Shape.zip", "data/polygons/wbd/HUC8_02060005.zip")
unzip(zipfile = "data/polygons/wbd/HUC8_02060005.zip", 
      exdir = "data/polygons/wbd")
```

HUC 8s - Read in USGS defined HUC 8 boundaries for Tuckahoe and Greensboro, merge into one polygon. 

```{r}
wbd <- st_read("data/polygons/wbd/Shape/WBDHU10.shp") %>% 
  dplyr::filter(Name %in% c("Tuckahoe Creek", "Watts Creek-Choptank River"))
ucrw <- wbd %>% st_union()
plot(ucrw)
# st_write(ucrw, "data/UCRW.shp")
```

HUC 12s - Read in USGS defined HUC 12 boundaries 

```{r}
huc12 <- st_read("data/polygons/wbd/Shape/WBDHU12.shp")
tuckahoe <- huc12 %>% 
  filter(HUC12 %in% grep("0206000501", huc12$HUC12, value = TRUE))
# plot(tuckahoe$geometry)
watts <- huc12 %>% 
  filter(HUC12 %in% grep("0206000502", huc12$HUC12, value = TRUE))
# plot(watts$geometry)
# plot(tuckahoe$geometry, add = TRUE)
ucrwHUC12 <- rbind(tuckahoe, watts)
# plot(ucrwHUC12$geometry)
```

```{r}
ucrwHUC12 %>%
  leaflet() %>%
  # addTiles() %>%
  addProviderTiles(providers$Esri.WorldImagery) %>%
  addPolygons(label = ~Name)
```

Subset to the top 4 headwater catchments and join them as one polygon

```{r}
AOI <- ucrwHUC12 %>% 
  dplyr::filter(Name %in% c("Cow Marsh Creek", "Upper Mason Branch", 
                            "Tappahanna Ditch-Choptank River", "Gravelly Branch-Choptank River"))

AOI <- AOI %>% st_union(by_feature = FALSE)
# plot(AOI_union)
# plot(AOI$geometry)
st_write(AOI, "data/polygons/AOI.shp")
st_write(AOI_union, "data/polygons/AOI_union.shp")
```

```{r}
AOI %>% 
  leaflet() %>%
  # addTiles() %>%
  addProviderTiles(providers$Esri.WorldImagery) %>%
  addPolygons(label = ~Name)
```

# MVdeps 

Use dataset from M Vanderhoof (link)[https://www.sciencebase.gov/catalog/item/59370e80e4b0f6c2d0d899d5]

Depression IDs generated here based on entire dataset

There are 3 files

* depressions
* Landsat 2015
* TotalStreamNetwork

```{r}
mv_dir <- "/nfs/khondula-data/Delmarva/data/vanderhoofUCRW"
# landsat <- sf::st_read(mv_dir, "Landsat_2015_UCRW") %>% st_as_sf()
mvdepressions <- sf::st_read(mv_dir, "depressions_UCRW") %>% 
  st_as_sf() 
mvdepressions <- mvdepressions %>% 
  mutate(dep_id = str_pad(1:nrow(mvdepressions), width = 5, pad = "0"))
st_write(mvdepressions, "data/polygons/mvdepressions.shp")
```

```{r}
mvdepressions <- sf::st_read("data/polygons/mvdepressions.shp")
head(mvdepressions)
mvdepressions_df <- mvdepressions %>% st_drop_geometry() %>%
  dplyr::mutate(area_m2 = ha*10000)
```

Leaflet map of all depressions

```{r, eval=FALSE}
# mvdepressions %>% 
#   st_transform(4326) %>%
#   leaflet() %>%
#   setView(lng = -75.825057, lat = 39.061090, zoom = 14) %>%
#   addProviderTiles(providers$Esri.WorldImagery) %>%
#   # addTiles() %>%
#   addPolygons()
```

Leaflet map of all depressions along with WBD 

```{r, eval=FALSE}
# wbd <- "https://hydro.nationalmap.gov/arcgis/services/wbd/MapServer/WMSServer"
# 
# mvdepressions %>% 
#   st_transform(4326) %>%
#   leaflet() %>%
#   setView(lng = -75.825057, lat = 39.061090, zoom = 14) %>%
#   addProviderTiles(providers$Esri.WorldImagery) %>%
#   # addTiles() %>%
#   addPolygons(group = "depressions") %>%
#   addWMSTiles(wbd, layers = "7",
#               options = WMSTileOptions(format = "image/png", transparent = TRUE),
#               group = "WBD transparent") %>%
#   addLayersControl(overlayGroups = c("depressions", "WBD transparent"))
```

## Size distribution of depressions

```{r}
mvdepressions %>%
  ggplot(aes(x = ha)) +
  geom_histogram() +
  scale_x_log10()
```

```{r}
mvdepressions %>%
  ggplot(aes(x = ha)) +
  stat_ecdf() +
  scale_x_log10() +
  theme_bw() + 
  ylab("Proportion") +
  xlab("Size (ha)")
```

Clip everything below some size threshold for faster plotting

```{r, eval = FALSE}
# wbd <- "https://hydro.nationalmap.gov/arcgis/services/wbd/MapServer/WMSServer"
# 
# mvdepressions %>% 
#   dplyr::filter(ha > 0.5) %>%
#   st_transform(4326) %>%
#   leaflet() %>%
#   setView(lng = -75.825057, lat = 39.061090, zoom = 14) %>%
#   addProviderTiles(providers$Esri.WorldImagery) %>%
#   # addTiles() %>%
#   addPolygons(group = "depressions", popup = ~as.character(dep_id)) %>%
#   # addPolygons(data = st_transform(ucrw, 4326)) %>%
#   # addWMSTiles(wbd, layers = "7",
#   #             options = WMSTileOptions(format = "image/png", transparent = TRUE),
#   #             group = "WBD transparent") %>%
#   addLayersControl(overlayGroups = c("depressions"))
```


# Crop depressions to AOI


```{r}
planet_crs <- "+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
```

```{r}
mvdeps_cropAOI <- mvdepressions %>% 
  st_transform(planet_crs) %>% 
  st_crop(st_transform(AOI_union, planet_crs))
```

Buffer to 20m 

```{r}
mvdeps_cropAOI_buffer20m <- mvdeps_cropAOI %>% 
  st_transform(5070) %>% 
  st_buffer(20) %>% 
  st_transform(planet_crs)

st_write(mvdeps_cropAOI_buffer20m,
         "/nfs/khondula-data/planetmethane/polygons/mvdeps_cropAOI_buffer20m.shp")
```

# Subset depressions

Use NWI dataset to identify overlap with farm ponds and forested wetlands.

Read in NWI and crop to AOI extent

```{r}
# download.file(url = "http://www.fws.gov/wetlands/downloads/Watershed/HU8_02060005_watershed.zip", "/nfs/khondula-data/planetmethane/nwi/HU8_02060005_watershed.zip")
# unzip("/nfs/khondula-data/planetmethane/nwi/HU8_02060005_watershed.zip",
#       exdir = "/nfs/khondula-data/planetmethane/nwi/")
nwi <- sf::st_read("/nfs/khondula-data/planetmethane/nwi/HU8_02060005_Watershed",
layer = "HU8_02060005_Wetlands")
nwi_prj <- nwi %>% st_transform(crs = planet_crs)

nwi_cropAOI <- nwi_prj %>% st_crop(AOI_union)
# plot(nwi_cropAOI$geometry)

# choptank_ws <- sf::st_read("/nfs/khondula-data/planetmethane/nwi/HU8_02060005_Watershed",
# layer = "HU8_02060005_Watershed")
# choptank_ws_prj <- choptank_ws %>% 
#   st_transform(crs = planet_crs)

```

Training data 

```{r}
AOI <- st_read("data/polygons/AOI_union.shp") %>% st_transform(planet_crs)
ww_areas_maskNWImvdeps <- raster::stack("data/ww_areas_maskNWImvdeps.tif")

nwi_FPLAKE <- st_read(file.path(data_dir, "nwi", "HU8_02060005_Watershed", "HU8_02060005_Wetlands.shp")) %>% 
                 dplyr::filter(WETLAND_TY %in% c("Freshwater Pond", "Lake")) %>%
                 st_transform(planet_crs)
nwi_FPLAKE_crop <- st_join(nwi_FPLAKE, AOI, left = FALSE)
# nwi_FPLAKE_crop %>% st_write("/nfs/khondula-data/planetmethane/polygons/nwi_FPLAKE_cropAOI.shp")
# plot(nwi_FPLAKE_crop$geometry)
```

```{r}
# library(leaflet)
# nwi_FPLAKE_crop %>%
#   st_transform(4326) %>%
#   leaflet() %>%
#   addProviderTiles(providers$Esri.WorldImagery) %>%
#   addRasterImage(ww_areas_maskNWImvdeps) %>%
#   addPolygons(opacity = 1, fillColor = "white", color = "white", weight = 1, group = "nwi") %>%
#   addLayersControl(overlayGroups = c("nwi"))
```


```{r}
  r_train_upland <- ww_areas_maskNWImvdeps %>% rasterToPolygons(n = 16, dissolve = TRUE)
  r_train_upland <- r_train_upland %>% st_as_sf() %>% 
    st_transform(planet_crs) %>% dplyr::mutate(trainclass = "upland")
  r_train_upland$trainid <- paste0("upland", 1:nrow(r_train_upland))
  r_train_upland <- r_train_upland[,-1]

  train_water <- nwi_FPLAKE_crop %>% st_transform(planet_crs) %>% dplyr::mutate(trainclass = "water")
  train_water$trainid <- paste0("water", 1:nrow(train_water))
  train_water <- train_water[,-c(1:6)]

  training_polygons_sf <- rbind(r_train_upland, train_water)

# training_polygons_sf %>% st_write("/nfs/khondula-data/planetmethane/polygons/training_polygons_v2.shp")
```


Take out farm ponds

```{r}
farmed_nwi <- dplyr::filter(nwi_cropAOI, ATTRIBUTE == "Pf")

mat <- mvdeps_cropAOI %>% st_intersects(farmed_nwi, sparse = FALSE)
depressions_pF <- which(apply(mat, 1, any))

mvdeps_cropAOI_subset <- mvdeps_cropAOI[-depressions_pF,]
```
