==> 03-extract-vals.Rmd <==
---
title: "Extract training"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(raster)
library(velox)
library(tidyr)
library(sf)
library(dplyr)
library(tabularaster)
library(fasterize)
library(fs)
library(sp)
library(readr)
data_dir <- "data"
planet_crs <- "+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
```

Images are here
```{r}
# band_names <- c("b1_blue", "b2_green", "b3_red", "b4_NIR",
#                 "NDVI", "NDWI", "HUE", "SAT", "VAL", "LUM", "CHROMA", "HUE2")
masked_layer_files <- fs::dir_ls("data/imgs_for_analysis")
masked_layer_files_nfs <- gsub("data", "/nfs/khondula-data/planetmethane", masked_layer_files)
```


```{r}
# masked_layer_files <- fs::dir_ls("data/imgs_for_analysis")
# imgfile <- masked_layer_files[1]
# imgfile <- allbands_notextracted[2]

extract_allbands <- function(imgfile){
  data_dir <- "/nfs/khondula-data/planetmethane"
  planet_crs <- "+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  band_names <- c("b1_blue", "b2_green", "b3_red", "b4_NIR",
                "NDVI", "NDWI", "HUE", "SAT", "VAL", "LUM", "CHROMA", "HUE2")
  img_extract_dir <- file.path(data_dir, "results", "masked-allbands-extract_spIDs")
  if(!fs::dir_exists(img_extract_dir)){fs::dir_create(img_extract_dir)}
  cells_extract_dir <- file.path(data_dir, "results", "masked-allbands-extract_cellids")
  if(!fs::dir_exists(cells_extract_dir)){fs::dir_create(cells_extract_dir)}

  # read in image
  imgstack <- imgfile %>% raster::stack()
  thisimgID <- substr(basename(imgfile), 1, 20)
  names(imgstack) <- band_names

  # training data
  training_polygons_sf <- st_read("/nfs/khondula-data/planetmethane/polygons/training_polygons_v2.shp")
  training_polygons_water <- dplyr::filter(training_polygons_sf, trainclass == "water")
  training_polygons_upland <- dplyr::filter(training_polygons_sf, trainclass == "upland")
  
  training_polygons_sp <- sf::as_Spatial(st_geometry(training_polygons_sf),
                  IDs = as.character(training_polygons_sf[["trainid"]])) 
  training_polygons_water_sp <- sf::as_Spatial(st_geometry(training_polygons_water),
                  IDs = as.character(training_polygons_water[["trainid"]]))
  training_polygons_uplan_sp <- sf::as_Spatial(st_geometry(training_polygons_upland),
                  IDs = as.character(training_polygons_upland[["trainid"]]))
  
  mvdeps_polygons_sf <- st_read("/nfs/khondula-data/planetmethane/polygons/mvdeps_buff20_inNLCD.shp")
  mvdeps_polygons_sp <- sf::as_Spatial(st_geometry(mvdeps_polygons_sf),
                    IDs = as.character(mvdeps_polygons_sf[["dep_id"]]))
 

  # get data frame of values with cell ids using tabularaster
  imgstack_df <- tabularaster::as_tibble(imgstack)
  imgstack_df_spread <- imgstack_df %>% spread(dimindex, cellvalue)
  names(imgstack_df_spread)[2:13] <- band_names
  imgstack_df_spread_narm <- dplyr::filter(imgstack_df_spread, !is.na(b1_blue))

  #
  imgstack_cell_water <- cellnumbers(imgstack, query = training_polygons_water_sp) 
  imgstack_cell_upland <- cellnumbers(imgstack, query = training_polygons_uplan_sp) 
  imgstack_cell_mvdeps <- cellnumbers(imgstack, query = mvdeps_polygons_sp) 

  imgstack_df_spread_wIDs <- imgstack_df_spread_narm %>% 
    mutate(water_training = cellindex %in% imgstack_cell_water$cell_,
           upland_training = cellindex %in% imgstack_cell_upland$cell_,
           mvdeps = cellindex %in% imgstack_cell_mvdeps$cell_)
  imgstack_df_spread_wIDs$imgID <- thisimgID
  imgstack_xys <- raster::xyFromCell(imgstack, imgstack_df_spread_wIDs$cellindex) 
  imgstack_df_spread_wIDs$x <- imgstack_xys[,1]
  imgstack_df_spread_wIDs$y <- imgstack_xys[,2]
  
  extract1_filename <- file.path(cells_extract_dir, sprintf("%s.csv", thisimgID))
  readr::write_csv(imgstack_df_spread_wIDs, extract1_filename)

  # because tabularaster doesnt keep sp IDs
  # do value extract with velox
  img_vx <- velox::velox(imgstack)
  img_extract <- img_vx$extract(training_polygons_sp, df = TRUE)
  names(img_extract)[2:13] <- band_names
  img_extract_narm <- img_extract %>% dplyr::filter(!is.na(NDWI)) %>% mutate(imgID = thisimgID)
  img_extract_narm <- img_extract_narm %>% mutate(trainclass = substr(ID_sp, 1, 5))

  extract2_filename <- file.path(img_extract_dir, sprintf("%s.csv", thisimgID))
  readr::write_csv(img_extract_narm, extract2_filename)
}

head(img_extract_narm)

# img_extract_narm$trainclass %>% table()
```

```{r}
# extract_allbands(imgfile = masked_layer_files[1])
```

```{r}
# testdf <- readr::read_csv(fs::dir_ls("/nfs/khondula-data/planetmethane/results/masked-allbands-extract/")[2])
# head(testdf)
```

```{r}
# masked_layer_files
# extract_allbands(masked_layer_files[2])

masked_layer_files <- fs::dir_ls("data/imgs_for_analysis")
masked_layer_ids <- substr(basename(masked_layer_files), 1, 20)
message(paste(length(masked_layer_ids), "masked images with bands added and without metadata issues"))

allbands_extracted <- substr(basename(fs::dir_ls("data/results/masked-allbands-extract_cellIDs/")), 1, 20)
message(paste(length(allbands_extracted), "cell extractions"))

allbands_extracted <- substr(basename(fs::dir_ls("data/results/masked-allbands-extract_spIDs/")), 1, 20)
message(paste(length(allbands_extracted), "polygon extractions"))

allbands_notextracted <- masked_layer_files[!masked_layer_ids %in% allbands_extracted]
# allbands_notextracted
message(paste(length(allbands_notextracted), "images with bands added but not extracted"))
```

```{r}
extract_allbands(allbands_notextracted[7])
purrr::walk(allbands_notextracted[1:9], ~extract_allbands(.x))
purrr::walk(masked_layer_files[21:40], ~extract_allbands(.x))
purrr::walk(masked_layer_files[41:60], ~extract_allbands(.x))
purrr::walk(masked_layer_files[61:80], ~extract_allbands(.x))
purrr::walk(masked_layer_files[81:100], ~extract_allbands(.x))
purrr::walk(masked_layer_files[101:113], ~extract_allbands(.x))
```

Or run on cluster
```{r}
library(rslurm)
allbands_notextracted_nfs <- gsub("data", "/nfs/khondula-data/planetmethane", allbands_notextracted)
pars <- data.frame(imgfile = allbands_notextracted_nfs,
                   stringsAsFactors = FALSE)

sjob5 <- rslurm::slurm_apply(extract_allbands, pars, 
                            jobname = "extract5", 
                            pkgs = c("tabularaster", "fasterize", "velox", "raster", "sf", "dplyr", "readr", "tidyr", "fs"),
                            slurm_options = list(time = "24:00:00", partition = "sesync"),
                            nodes = 5, 
                            cpus_per_node = 4,
                            submit = TRUE)

# rslurm::print_job_status(sjob)
# rslurm::print_job_status(sjob2)
# rslurm::cancel_slurm(sjob5)
rslurm::print_job_status(sjob5)
length(list.files("data/results/masked-allbands-extract_cellIDs"))
length(list.files("data/results/masked-allbands-extract_spIDs"))
```

```{r}
orig_img_ids <- list.files("data/data/") %>% substr(1,20)
masked_imgfiles <- fs::dir_ls("data/udm_masked_imgs/", glob = "*.grd")
udm_masked_ids <- substr(basename(masked_imgfiles), 1, 20)
addbands_files <- fs::dir_ls("data/imgs_for_analysis")
addbands_ids <- substr(basename(addbands_files), 1, 20)
img_metadata_df <- readr::read_csv("data/img_metadata_df.csv")

data.frame(planet_downloads = orig_img_ids,
           udm_mask = orig_img_ids %in% udm_masked_ids,
           bands_added = orig_img_ids %in% addbands_ids) %>%
  left_join(img_metadata_df, by = c("planet_downloads" = "imgID")) %>%
  write_csv("images_inventory.csv")
```

Function to get cell number to mvdeps polygon IDs for each image

```{r}

# imgID # mvdepsID # cellindex
# imgfile <- masked_layer_files[7]
# img_id <- "20171221_151120_1035" # good

library(raster)
library(sf)
library(dplyr)
library(readr)
library(tabularaster)
library(fs)
# library(parallel)

get_mvdeps_cellnos <- function(imgfile, my_dep_id = "07960"){
  data_dir <- "/nfs/khondula-data/planetmethane"
  planet_crs <- "+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs +ellps=WGS84 +towgs84=0,0,0"
  band_names <- c("b1_blue", "b2_green", "b3_red", "b4_NIR",
                "NDVI", "NDWI", "HUE", "SAT", "VAL", "LUM", "CHROMA", "HUE2")

   imgstack <- imgfile %>% raster::stack()
    thisimgID <- substr(basename(imgfile), 1, 20)
    names(imgstack) <- band_names
    imgstack_onelayer <- imgstack[[1]]
    mvdeps_polygons_sf <- sf::st_read("/nfs/khondula-data/planetmethane/polygons/mvdeps_buff20_inNLCD.shp")
    mvdeps_polygons_sf <- st_transform(mvdeps_polygons_sf, planet_crs)

   # cellnos_list <- mvdeps_polygons_sf$dep_id %>% as.character() %>%
     # purrr::map(~tabularaster::cellnumbers(imgstack_onelayer, 
                                           # query = filter(mvdeps_polygons_sf, dep_id == .x)))
   
    cellnos_df <- tabularaster::cellnumbers(imgstack_onelayer, query = filter(mvdeps_polygons_sf, dep_id == my_dep_id))

   # names(cellnos_list) <- mvdeps_polygons_sf$dep_id %>% as.character()
   # imgstack_cell_mvdeps_sf <- bind_rows(cellnos_list, .id = "ID_sp")

    this_dir <- file.path(data_dir, "metadata", "mvdeps_cellnos2", thisimgID)
    if(!fs::dir_exists(this_dir)){fs::dir_create(this_dir)}

      mvdeps_cellno_filename <- file.path(this_dir, sprintf("dep_id_%s.csv", my_dep_id))
      cellnos_df %>% write_csv(mvdeps_cellno_filename)
}


```

```{r}
# mvdeps_polygons_sf <- sf::st_read("data/polygons/mvdeps_buff20_inNLCD.shp")
# mvdep_ids <- mvdeps_polygons_sf$dep_id %>% as.character()
# purrr::walk(mvdep_ids[1:100], ~get_mvdeps_cellnos(imgfile = masked_layer_files[7], my_dep_id = .x))
# purrr::walk2(.x = rep(masked_layer_files[7], 100),
#              .y = mvdep_ids[1:100], ~get_mvdeps_cellnos(imgfile = .x, my_dep_id = .y))
# 
# img_cellnos <- fs::dir_ls("data/metadata/mvdeps_cellnos2/20171006_155635_104e/") %>%
#   map_df(~readr::read_csv(.x, col_types = c("dd")), .id = "filename")
```

Or run on cluster
```{r}
library(rslurm)
mvdeps_polygons_sf <- sf::st_read("data/polygons/mvdeps_buff20_inNLCD.shp")
mvdep_ids <- mvdeps_polygons_sf$dep_id %>% as.character()

pars <- data.frame(imgfile = rep(masked_layer_files_nfs, each = length(mvdep_ids)),
                   my_dep_id = rep(mvdep_ids, times = length(masked_layer_files)),
                   stringsAsFactors = FALSE)

sjob10 <- rslurm::slurm_apply(get_mvdeps_cellnos, pars, 
                            jobname = "cellnos2",
                            pkgs = c("tabularaster", "raster", "sf", "dplyr", "readr", "fs"),
                            slurm_options = list(partition = "sesync"),
                            nodes = 20, 
                            cpus_per_node = 8,
                            submit = TRUE)

# rslurm::cancel_slurm(sjob10)
rslurm::print_job_status(sjob10)
```

```{r}
length(list.files("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2"))
# length(list.files("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2/20171004_155646_103f/"))
# length(list.files("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2/20171020_155434_0f2e/"))
# length(list.files("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2/20171117_154719_104c/"))
# # how many files in each directory?
# list.files("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2", full.names = TRUE) %>% 
#   sapply(function(x) length(list.files(x)))

cellnos_progress <- fs::dir_ls("/nfs/khondula-data/planetmethane/metadata/mvdeps_cellnos2") %>%
  purrr::map_dbl(~length(list.files(.x)))

cellnos_progress_df <- data.frame(imgid = basename(names(cellnos_progress)),
                                  deps_files = cellnos_progress, 
                                  stringsAsFactors = FALSE,
                                  row.names = NULL) %>%
  arrange(desc(deps_files))
# how many complete?
complete_cellnos_imgs <- cellnos_progress_df %>% 
  filter(deps_files == 9290) %>% pull(imgid)
message(length(complete_cellnos_imgs), " imgs with cellnos extracted")

```

```{r}
# mvdeps_cellnos_dir <- file.path(data_dir, "metadata", "mvdeps_cellnos")
# mvdeps_cellno_ids <- fs::dir_ls(mvdeps_cellnos_dir) %>% basename() %>% substr(1,20)
# masked_layer_files <- fs::dir_ls("/nfs/khondula-data/planetmethane/imgs_for_analysis")
# masked_layer_ids <- substr(basename(masked_layer_files), 1, 20)
# 
# get_cellnos_imgs <- masked_layer_files[!masked_layer_ids %in% mvdeps_cellno_ids]
# message(length(get_cellnos_imgs), " images for analysis without mv deps cell numbers")
# # get_mvdeps_cellnos(imgfile = get_cellnos_imgs[2])

```

```{r}
# purrr::walk(get_cellnos_imgs, ~get_mvdeps_cellnos(.x))
```
